+++ src/tooling/piral-cli/src/common/config.ts
@@ -8,6 +8,7 @@ export interface PiralCliConfig {
   cert?: string;
   npmClient?: NpmClientType;
   bundler?: string;
+  piletApi?: string;
 }
 
 export const config: PiralCliConfig = rc('piral', {
@@ -37,5 +38,9 @@ export const config: PiralCliConfig = rc('piral', {
    * Selects the default bundler to use, if
    * none given and found.
    */
-  bundler: 'parcel',
+  bundler: 'webpack',
+  /**
+   * Selects the default pilet API path to use.
+   */
+  piletApi: '/$pilet-api',
 });